pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "geopy>=2.4.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pyswisseph>=2.10.3.2",
    "pytz>=2025.2",
    "uvicorn>=0.35.0",
]
